+++ modules/vivid-core/include/vivid/effects/effects.h
@@ -23,6 +23,7 @@
 #include <vivid/effects/level.h>
 #include <vivid/effects/lookup.h>
 #include <vivid/effects/crop.h>
+#include <vivid/effects/fit.h>
 #include <vivid/effects/tone_map.h>
 #include <vivid/effects/normal_map.h>
 #include <vivid/effects/mirror.h>